+++ neo4j/src/driver/io/bolt.rs
@@ -20,6 +20,7 @@ mod bolt5x1;
 mod bolt5x2;
 mod bolt5x3;
 mod bolt5x4;
+mod bolt5x6;
 mod bolt_state;
 mod chunk;
 mod handshake;
@@ -59,6 +60,7 @@ use bolt5x1::{Bolt5x1, Bolt5x1StructTranslator};
 use bolt5x2::{Bolt5x2, Bolt5x2StructTranslator};
 use bolt5x3::{Bolt5x3, Bolt5x3StructTranslator};
 use bolt5x4::{Bolt5x4, Bolt5x4StructTranslator};
+use bolt5x6::{Bolt5x6, Bolt5x6StructTranslator};
 use bolt_state::{BoltState, BoltStateTracker};
 use chunk::{Chunker, Dechunker};
 pub(crate) use handshake::{open, TcpConnector};
@@ -175,6 +177,7 @@ impl<RW: Read + Write> Bolt<RW> {
             data: BoltData::new(version, stream, socket, local_port, address),
             // [bolt-version-bump] search tag when changing bolt version support
             protocol: match version {
+                (5, 6) => Bolt5x6::<Bolt5x6StructTranslator>::default().into(),
                 (5, 4) => Bolt5x4::<Bolt5x4StructTranslator>::default().into(),
                 (5, 3) => Bolt5x3::<Bolt5x3StructTranslator>::default().into(),
                 (5, 2) => Bolt5x2::<Bolt5x2StructTranslator>::default().into(),
@@ -487,7 +490,8 @@ enum BoltProtocolVersion {
     V5x1(Bolt5x1<Bolt5x1StructTranslator>),
     V5x2(Bolt5x2<Bolt5x2StructTranslator>),
     V5x3(Bolt5x3<Bolt5x3StructTranslator>),
-    V5x4(Bolt5x4<Bolt5x3StructTranslator>),
+    V5x4(Bolt5x4<Bolt5x4StructTranslator>),
+    V5x6(Bolt5x6<Bolt5x6StructTranslator>),
 }
 
 #[derive(Debug, Copy, Clone, Ord, PartialOrd, Eq, PartialEq)]