+++ neo4j/src/driver/io/bolt/handshake.rs
@@ -34,9 +34,9 @@ use crate::time::Instant;
 const BOLT_MAGIC_PREAMBLE: [u8; 4] = [0x60, 0x60, 0xB0, 0x17];
 // [bolt-version-bump] search tag when changing bolt version support
 const BOLT_VERSION_OFFER: [u8; 16] = [
+    0, 0, 6, 5, // BOLT 5.6
     0, 4, 4, 5, // BOLT 5.4 - 5.0
     0, 0, 4, 4, // BOLT 4.4
-    0, 0, 0, 0, // -
     0, 0, 0, 0, // -
 ];
 
@@ -248,6 +248,7 @@ fn decode_version_offer(offer: &[u8; 4]) -> Result<(u8, u8)> {
         [0, 0, 0, 0] => Err(Neo4jError::InvalidConfig {
             message: String::from("server version not supported"),
         }),
+        [_, _, 6, 5] => Ok((5, 6)),
         [_, _, 4, 5] => Ok((5, 4)),
         [_, _, 3, 5] => Ok((5, 3)),
         [_, _, 2, 5] => Ok((5, 2)),
@@ -379,6 +380,7 @@ mod tests {
     #[case([0, 0, 2, 5], (5, 2))]
     #[case([0, 0, 3, 5], (5, 3))]
     #[case([0, 0, 4, 5], (5, 4))]
+    #[case([0, 0, 6, 5], (5, 6))]
     fn test_decode_version_offer(
         #[case] mut offer: [u8; 4],
         #[case] expected: (u8, u8),
@@ -418,7 +420,8 @@ mod tests {
     #[case([0, 0, 1, 4])] // driver didn't offer version 4.1
     #[case([0, 0, 2, 4])] // driver didn't offer version 4.2
     #[case([0, 0, 3, 4])] // driver didn't offer version 4.3
-    #[case([0, 0, 5, 5])] // driver didn't offer version 5.4
+    #[case([0, 0, 5, 5])] // driver didn't offer version 5.5
+    #[case([0, 0, 7, 5])] // driver didn't offer version 5.7
     #[case([0, 0, 0, 6])] // driver didn't offer version 6.0
     fn test_garbage_server_version(
         #[case] mut offer: [u8; 4],